requirements.txt
streamlit==1.19.0
watchdog==2.3.1
openai==0.27.0
python-decouple==3.8
altair<5